# Ninad 150096725180/docs/swagger.yaml
openapi: 3.0.3
info:
  title: 📚 Library Management System API
  version: 1.0.0
  description: |
    Complete RESTful API Assignment for Library Management built using Node.js, Express.js, JWT, bcrypt, and Firebase Firestore.
    Features Role-Based Access Control (Student vs Librarian), complete CRUD operations, book borrowing/returning, rate limiting, and structured logging.
  contact:
    name: Ninad Nilesh Deodhare
    email: student@library.com
servers:
  - url: http://localhost:5000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token obtained from /api/auth/login or /api/auth/register.
...
paths:
...
  /api/books:
    get:
      summary: Get all books
      description: Retrieve a list of books. Supports filtering by category, status, and author.
      tags:
        - Books
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter books by category (e.g. Technology, Fiction)
        - in: query
          name: status
          schema:
            type: string
            enum: [available, borrowed]
          description: Filter by status
        - in: query
          name: author
          schema:
            type: string
          description: Filter by exact author
      responses:
        200:
          description: List of books matching criteria

    post:
      summary: Add new book (Librarian only)
      description: Creates a new book entity in Firestore. Requires librarian credentials.
      tags:
        - Books
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - isbn
                - category
              properties:
                title:
                  type: string
                  example: "Refactoring: Improving the Design of Existing Code"
                author:
                  type: string
                  example: "Martin Fowler"
                isbn:
                  type: string
                  example: "978-0134757599"
                category:
                  type: string
                  example: "Technology"
                quantity:
                  type: integer
                  default: 1
                  example: 3
      responses:
        201:
          description: Book added successfully
        400:
          description: Validation error
        401:
          description: Unauthorized
        403:
          description: Forbidden - Librarian role required
...
  /api/transactions:
    get:
      summary: Get all transactions (Librarian only)
      description: Retrieves complete library borrowing transactions with user and book details.
      tags:
        - Transactions
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of transactions
        401:
          description: Unauthorized
        403:
          description: Forbidden (Librarian required)
...
  /api/users:
    get:
      summary: Get all users (Librarian only)
      tags:
        - User Management
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of all users (passwords omitted)
        401:
          description: Unauthorized
        403:
          description: Forbidden (Librarian required)
...